Add files via upload
Gest.Viff is not a commonly known file format in Next.js or web development. However, based on the name, it could be:

A custom font file used in your project.
A vector image file (similar to .vif, used in imaging software).
A configuration or asset file specific to a library or framework.